mktemp: move to Formula and compat.
Move this method away from FileUtils as it doesn't need to be on that
class.
